requirements.txt
aiogram>=3.4
asyncpg>=0.29
apscheduler>=3.10
python-dotenv>=1.0
